requirements.txt
# Python >= 3.10
google-genai==1.55.0
pydantic==2.12.5
python-dotenv==1.2.1
fastmcp==3.0.0
mcp==1.26.0
